+++ modules/contracts/src.ts/services/ethService.ts
@@ -15,7 +15,6 @@ import {
   TransactionEventMap,
   StringifiedTransactionReceipt,
   StringifiedTransactionResponse,
-  GAS_ESTIMATES,
   TransactionResponseWithResult,
   getConfirmationsForChain,
 } from "@connext/vector-types";
@@ -27,6 +26,7 @@ import {
   getRandomBytes32,
   hashCoreTransferState,
 } from "@connext/vector-utils";
+import { Interface } from "@ethersproject/abi";
 import { Signer } from "@ethersproject/abstract-signer";
 import { BigNumber } from "@ethersproject/bignumber";
 import { Contract } from "@ethersproject/contracts";
@@ -44,8 +44,8 @@ import { ChannelFactory, VectorChannel } from "../artifacts";
 
 import { EthereumChainReader } from "./ethReader";
 
-export const EXTRA_GAS_PRICE = parseUnits("20", "gwei");
 export const EXTRA_GAS = 50_000;
+export const BIG_GAS_LIMIT = BigNumber.from(1_000_000); // 1M gas should cover all Connext txs
 
 export const waitForTransaction = async (
   provider: JsonRpcProvider,
@@ -246,9 +246,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           if (multisigRes.getValue() !== `0x`) {
             return undefined;
           }
-          const _gas = GAS_ESTIMATES.createChannel;
-          const gas = _gas.add(EXTRA_GAS);
-          return channelFactory.createChannel(channelState.alice, channelState.bob, { gasPrice, gasLimit: gas });
+          return channelFactory.createChannel(channelState.alice, channelState.bob, { gasPrice, gasLimit: BIG_GAS_LIMIT });
         },
       );
       if (result.isError) {
@@ -310,12 +308,16 @@ export class EthereumChainService extends EthereumChainReader implements IVector
             return this.sendDepositATx(channelState, amount, AddressZero, gasPrice);
           }
           // otherwise deploy with deposit
-          const _gas = GAS_ESTIMATES.createChannelAndDepositAlice;
-          const gas = _gas.add(EXTRA_GAS);
+          const data = new Interface(ChannelFactory.abi).encodeFunctionData("createChannelAndDepositAlice", [
+            channelState.alice,
+            channelState.bob,
+            assetId,
+            amount,
+          ]);
           return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
             value: amount,
             gasPrice,
-            gasLimit: gas,
+            gasLimit: BIG_GAS_LIMIT,
           });
         },
       ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
@@ -355,11 +357,9 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           // multisig deployed, just send deposit (will check allowance)
           return this.sendDepositATx(channelState, amount, assetId, gasPrice);
         }
-        const _gas = GAS_ESTIMATES.createChannelAndDepositAlice;
-        const gas = _gas.add(EXTRA_GAS);
         return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
           gasPrice,
-          gasLimit: gas,
+          gasLimit: BIG_GAS_LIMIT,
         });
       },
     ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
@@ -392,8 +392,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     if (gasPriceRes.isError) {
       Result.fail(gasPriceRes.getError()!);
     }
-    const _gasPrice = gasPriceRes.getValue();
-    const gasPrice = _gasPrice.add(EXTRA_GAS_PRICE);
+    const gasPrice = gasPriceRes.getValue();
     this.log.info(
       {
         channelAddress: channelState.channelAddress,
@@ -459,19 +458,12 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     }
 
     this.log.info({ sender, method, methodId, channel: channelState.channelAddress }, "Sending withdraw tx to chain");
-    const gasEstimateRes = await this.estimateGas(channelState.networkContext.chainId, { ...minTx, from: sender });
-    if (gasEstimateRes.isError) {
-      return Result.fail(gasEstimateRes.getError()!);
-    }
-    const _gas = gasEstimateRes.getValue();
-    const gas = _gas.add(EXTRA_GAS);
-    this.log.info({ method, methodId, gas: gas.toString() }, "Estimated gas");
     return this.sendTxWithRetries(
       channelState.channelAddress,
       channelState.networkContext.chainId,
       TransactionReason.withdraw,
       async () => {
-        return signer.sendTransaction({ ...minTx, gasPrice, gasLimit: gas, from: sender });
+        return signer.sendTransaction({ ...minTx, gasPrice, gasLimit: BIG_GAS_LIMIT, from: sender });
       },
     ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
@@ -518,8 +510,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     if (gasPriceRes.isError) {
       Result.fail(gasPriceRes.getError()!);
     }
-    const _gasPrice = gasPriceRes.getValue();
-    const gasPrice = _gasPrice.add(EXTRA_GAS_PRICE);
+    const gasPrice = gasPriceRes.getValue();
     this.log.info(
       {
         channelAddress: channelState.channelAddress,